requirements/base.txt
Babel==2.4.0
Django==1.11.2
django-environ==0.4.3
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
MySQL-python==1.2.5
phonenumberslite==8.5.0
pytz==2017.2
six==1.10.0
